// mtrakr-bul/src/app/models/Volunteer.ts
export class Volunteer {
  id?: string;
  dateUnAvailable?: any;
  eventTypesAvailable?: string;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address1: string;
  address2: string;
  city: string;
  state: string;
  zipcode: number;
  familyID: string;
  isAvailable: boolean;
  isAdmin: boolean;
  isCantor: boolean;
  isEMoHC: boolean;
  isGifts: boolean;
  isGiftsChild: boolean;
  isLector: boolean;
  isMassCoord: boolean;
  isOther: boolean;
  isRosary: boolean;
  isServer: boolean;
  isTech: boolean;
  isUsher: boolean;
  isSaturday: boolean;
  isSundayEarly: boolean;
  isSundayLate: boolean;
  isWeekday: boolean;
}